Shift rerouting responsibility to RouteController
RouteController is now responsible for the basic aspects of rerouting, including fetching the new route.

A new RouteControllerDelegate protocol provides rerouting “lifecycle” methods so that a delegate such as NavigationViewController can preempt rerouting and perform additional rerouting tasks like updating UI. Since the delegate pattern is one-to-one, any other classes that need to respond to rerouting can still observe notifications about rerouting. However, those other classes cannot preempt rerouting as the delegate can. NavigationViewController itself now gives its own delegate the chance to preempt rerouting, which is consistent with the readme.

Rationalized the rerouting notification to reflect the fact that the decision has already been made to reroute by the time the notification is posted. Renamed the key in the notification’s user info dictionary to describe the key’s associated value instead of the notification that includes the key.

When rerouting, any remaining intermediate waypoints are preserved in the new route, not only the ultimate destination.