Suppress freeway names in post-maneuver instructions (#412)
Refactored the post-maneuver instruction generation code to account for the road class and moved the code to an extension on RouteStep. Consistently mark up road names and route numbers as addresses when either (but not both) is available on the step in question. In the rare event that a motorway has no name or number, deliver a post-maneuver instruction without the “on [road name]” prepositional phrase.