Force-update showsUserLocation
showsUserLocation preinitializes the user location annotation view very early on, but the delegate has yet to be set by that time, let alone the navigation map delegate. When each of the delegate properties is set, force the map view to reconsider everything it does as part of showing the user location.