Fix lane arrows at forks
If the maneuver direction appears to contradict the lane indication by going to the opposite side, it’s probably because of a fork in which a lane can be used to take either branch. In that case, ignore the invalid branch and show only the valid branch, going in the same direction as the maneuver. Ideally, we’d show the invalid branch dimmed out, but we currently lack artwork for a left-right turn lane.