Attributed distance strings
Implemented DistanceFormatter.attributedString(for:), which applies a “quantity” attribute to the numeric quantity part of the returned attributed string. This method relies on finding the formatted number within the distance string, as opposed to finding the unit, which turned out to be unreliable in some locales.

Simplified BaseInstructionsBannerView.distance’s setter to pass this attributed string directly to DistanceLabel instead of attempting to split the distance string on its quantity and unit. DistanceLabel now emphasizes the quantity relative to the unit based on the presence of the quantity attribute.

Added tests of attributed distance formatting.