Test-drive CarPlayManager into telling its delegate when a trip begins
Not sure if these testing patterns are going to be sustainable, but
underlying CarPlay functionality gets stuck in semaphore wait traps, and
many CarPlay objects cannot be created directly. Might need to lean on a
test doubles framework to make this better. Good news is these are all
Objective-C classes, so runtime indirection is possible for unit tests.