Pan based on screen distance in CarPlay
The panning buttons now shift the map by a fixed number of screen points rather than a fixed physical distance. The previous approach resulted in wildly different behavior depending on the zoom level and relied on the map to be unrotated, which isn’t quite tenable given the recent change to follow with course.