Allow style customization in CarPlay
Renamed CarPlayManager and CarPlayNavigationViewController initializers to accept Style objects, which are applied throughout the CarPlay interface, similar to NavigationViewController. DayStyle and NightStyle are still used by default but no longer hard-coded everywhere.

Renamed the StyleManager initializer to rely on a delegate property instead of an argument that looks like it gives ownership of the delegate to StyleManager.

Exposed a few more CarPlayManager members to Objective-C code.