Stop linting MapboxNavigation.podspec
MapboxNavigation fails the linter between releases because it depends on MapboxCoreNavigation. The linter installs the latest release of MapboxCoreNavigation, which potentially has a different set of public symbols than what’s in the current branch. Moreover, at release time, it fails the linter because the new release of MapboxCoreNavigation hasn’t been published yet.