Inlined conversion from Route to JSON string
Conversion to a JSON string is a very specialized operation that we only need in one place for working with MapboxNavigationNative. This operation is likely too expensive to justify a computed variable. The computed variable was sitting in its own file, creating a naming conflict that broke the jazzy docset.