Fixed Xcode 11.4 builds
Switched to a branch of SnappyShrimp that supports building on Xcode 11.4. Build this configuration on CircleCI.

Renamed snapshot test fixtures for iOS 13.4. When simulating navigation in CarPlay, start the navigator directly instead of going through the CarPlayNavigationViewController, which goes away prematurely in iOS 13.4 because there’s no view hierarchy to add it to.