Fixed CarPlayNavigationViewController tests
Tests that exercise CarPlayNavigationViewController need it to be around after being presented. Presenting it requires the CarPlayMapViewController to be part of the view hierarchy. Merely making CarPlayMapViewController the root view controller of the CPWindow doesn’t automatically add the view to the view hierarchy, though it does happen in a realistic, non-testing scenario.