Fixed testSnappedAtEndOfStepLocationWhenCourseIsSimilar
Previously, the test attempted to simulate making a wrong turn by negating the expected final heading and using that as the course of the location update. But a negative course is an invalid course that we avoid passing along to the navigator. The correct way to turn in the opposite direction is to add 180°.