Unset monotonic timestamp
The idea behind introducing the monotonicTimestampNanoseconds parameter to FixLocation was apparently that the navigator could rely on a guaranteed monotonic value to sort locations and detect and discard any jumps in time that may be caused by changes to or unreliability in the system time. The timestamp in each location update serves a somewhat different purpose than a monotonic counter, since it corresponds specifically to the time at which the system thinks it received the location update. A significant amount of time may have elapsed since the location update, and that gap may not be consistent between location updates, a fact that would be obscured by relying on the monotonic counter instead of the location timestamp. To avoid misunderstandings about the origin and behavior of the timestamp, this change goes back to specifying only a conventional timestamp and leaving the monotonic counter unspecified.