Restructured free driving classes
Inverted FreeDriveLocationManager to be a delegate of a CLLocationManager that is responsible for talking to the navigator in much the same way that RouteController does. Removed FreeDriveDebugInfoListener and FreeDriveLocationManager.ProxyDelegate in favor of FreeDriveLocationManagerDelegate.

Moved debugging utilities from MapboxNavigation to the example application targets and out of the project root. Removed remaining vestiges of MapboxNavigationNative from the MapboxNavigation public API.