Renamed free driving to passive location
The distinction between the MapboxCoreNavigation and MapboxNavigation free driving classes is easier to make if we call one a location data source, borrowing from RouteDataSource terminology, and the other a location manager, for consistency with MGLLocationManager. This naming convention also avoids the awkwardness of beginning class names with a noun phrase that could easily be misread as a verb phrase.

Added a changelog entry.