Partially implement MapboxNavigationTests in SPM
- All assets moved to TestHelper target for easier maintenance.
- Introduced a new test helper target - "CarPlayTestHelper", which holds
some mocks required for testing CarPlay. It had to be written in ObjC
because Swift forbids specific unsafe patterns. In particular:
CPNavigationSession and CPInterfaceController init methods are marked
unavailable.
- Use https://github.com/alanzeino/ios-snapshot-test-case.git in SPM,
although, it is likely to be replaced with https://github.com/pointfreeco/swift-snapshot-testing
which doesn't have all the features but has first-class Xcode and SPM
support. Or extended with a custom fork to support more configuration
points.
- All snapshot-based tests are disabled because it is not clear how to
provide configuration for https://github.com/alanzeino/ios-snapshot-test-case.git
version with SPM.
- Xcode base MapboxNavigationTests are not tested and should be fixed
separately.