Fixes the race in ImageDownload
When cancel is called `start` method can be executed on a different thread. There is a room when cancel hasn't updated the state yet and
`start` method verified it could start.
With the fix, we make sure that start can't be executed while cancel is
by guarding methods in a lock.