Fix obtaining latest tileset descriptor
The latest tile descriptor is resolved asynchronously in
`MBNNCacheHandle,` but there is no way to wait until the latest
descriptor is resolved in the cache handle. Until we have such
capabilities, we deploy a quick fix to let `MBNNCacheHandle` resolve
the latest descriptor by waiting for  X seconds until we ask it for
the descriptor. We fallback to this workaround only when `Navigator`
hasn't been created yet, or the navigator isn't in the appropriate state
for requested tile descriptor, otherwise we reuse Navigator's
cacheHandle.

The only change in TileStore.swift is adapting to the new async nature
of the `TilesetDescriptorFactory.getLatest` method.