SDK now uses global Directions and TileStoreConfiguration instances
Issue: #1039

With this change, it is clear that `TileStoreConfiguration` should be the
same for the app's lifetime. Providing different configurations is undefined behavior.

Also, it is possible now to override the default `Directions.shared` used in
various places of the app.

The primary reason for this refactoring is to simplify the `TilesetDescriptorFactory` API, which doesn't require creating and managing multiple `CacheHandle` lifetimes.