Fix routeProgress get out of sync with NavNative status
- `NavigationService.updateRoute`, `RouteController.updateRoute` is no
longer synchronous. This change is needed because NavNative
implementation is asynchronous, and we need to know when this operation
completes so that we can update local properties with a new route.
Otherwise, we can end up in a situation when the SDK state doesn't match
with the nav status callback received from NavNative.

- `userIsOnRoute(_:status:)` incorrectly assumes that the invalid route
state is an off-route event. The invalid route state is reported when
there is no route set in a native navigator, that happens during passive
navigation.

- It is no longer possible to set `RouteController.routeProgress`
property synchronously for the same reason as why
`RouteController.updateRoute` is asynchronous.

- `updateNavigator` incorrectly referenced old `routeProgress` in the
parameters to NavNative's `setRoute` method.

- `RouteProgress.legIndexHandler` removed as it is no longer needed.
`RouteProgress.legIndex` should be updated only after NavNative status
callback, otherwise they will be out-of-sync.
`RouteController.advanceLegIndex` is the correct method to change
legIndex, then it will be updated in `RouteProgress`.

- `NavigationEventsManagerTestDoubles.flush` method was adjusted to
work closer to real world. Without clearing enqueuedEvents, if flus is
called two times, flushed events will contain two duplicates.

- A lot of `RouteController` dependent tests are fixed, some restored.