Adds more flexibility to CarPlay route cancellation configuration
Introduce new `CarPlayManagerDelegate` delegate method
`CarPlayManagerDelegate.carPlayManagerWillCancelPreview(_:configuration:)`
to allow for customization of route cancellation behavior.

If `configuration.popToRoot` is `true`, the `popToRootTemplate` method
is called to navigate to the root template. This change enhances the
flexibility of route cancellation in CarPlay and is future proof by
allowing to add more options.